requirements.txt
# requirements.txt
numpy>=1.24
pandas>=2.2
scikit-learn>=1.4
xgboost>=2.0
lightgbm>=4.2
matplotlib>=3.9
python-dateutil>=2.9
requests>=2.32
holidays>=0.34
statsmodels>=0.14
joblib>=1.4
